[odoo] Updates till https://github.com/odoo/odoo/commit/bfaab1e7aa1ebe2f9cfc98ef38b114f30ec19e88

+++ addons/stock/models/stock_move_line.py
@@ -13,6 +13,7 @@ from flectra.tools.float_utils import float_round, float_compare, float_is_zero
 class StockMoveLine(models.Model):
     _name = "stock.move.line"
     _description = "Packing Operation"
+    _rec_name = "product_id"
     _order = "result_package_id desc, id"
 
     picking_id = fields.Many2one(
@@ -49,6 +50,7 @@ class StockMoveLine(models.Model):
     consume_line_ids = fields.Many2many('stock.move.line', 'stock_move_line_consume_rel', 'consume_line_id', 'produce_line_id', help="Technical link to see who consumed what. ")
     produce_line_ids = fields.Many2many('stock.move.line', 'stock_move_line_consume_rel', 'produce_line_id', 'consume_line_id', help="Technical link to see which line was produced with this. ")
     reference = fields.Char(related='move_id.reference', store=True)
+    in_entire_package = fields.Boolean(compute='_compute_in_entire_package')
 
     @api.one
     def _compute_location_description(self):
@@ -77,6 +79,13 @@ class StockMoveLine(models.Model):
         for `product_qty`, where the same write should set the `product_uom_qty` field instead, in order to
         detect errors. """
         raise UserError(_('The requested operation cannot be processed because of a programming error setting the `product_qty` field instead of the `product_uom_qty`.'))
+
+    def _compute_in_entire_package(self):
+        """ This method check if the move line is in an entire pack shown in the picking."""
+        for ml in self:
+            picking_id = ml.picking_id
+            ml.in_entire_package = picking_id and picking_id.picking_type_entire_packs and picking_id.state != 'done'\
+                                   and ml.result_package_id and ml.result_package_id in picking_id.entire_package_ids
 
     @api.constrains('product_uom_qty')
     def check_reserved_done_quantity(self):
@@ -339,7 +348,11 @@ class StockMoveLine(models.Model):
             if ml.product_id.type == 'product' and not ml.location_id.should_bypass_reservation() and not float_is_zero(ml.product_qty, precision_digits=precision):
                 self.env['stock.quant']._update_reserved_quantity(ml.product_id, ml.location_id, -ml.product_qty, lot_id=ml.lot_id,
                                                                    package_id=ml.package_id, owner_id=ml.owner_id, strict=True)
-        return super(StockMoveLine, self).unlink()
+        moves = self.mapped('move_id')
+        res = super(StockMoveLine, self).unlink()
+        if moves:
+            moves._recompute_state()
+        return res
 
     def _action_done(self):
         """ This method is called during a move's `action_done`. It'll actually move a quant from
@@ -426,9 +439,9 @@ class StockMoveLine(models.Model):
         if 'lot_id' in vals and vals['lot_id'] != move.lot_id.id:
             data['lot_name'] = self.env['stock.production.lot'].browse(vals.get('lot_id')).name
         if 'location_id' in vals:
-            data['location_name'] = self.env['stock.location_id'].browse(vals.get('location_id')).name
+            data['location_name'] = self.env['stock.location'].browse(vals.get('location_id')).name
         if 'location_dest_id' in vals:
-            data['location_dest_name'] = self.env['stock.location_id'].browse(vals.get('location_dest_id')).name
+            data['location_dest_name'] = self.env['stock.location'].browse(vals.get('location_dest_id')).name
         if 'package_id' in vals and vals['package_id'] != move.package_id.id:
             data['package_name'] = self.env['stock.quant.package'].browse(vals.get('package_id')).name
         if 'package_result_id' in vals and vals['package_result_id'] != move.package_result_id.id:
@@ -460,7 +473,6 @@ class StockMoveLine(models.Model):
                 ('owner_id', '=', owner_id.id if owner_id else False),
                 ('package_id', '=', package_id.id if package_id else False),
                 ('product_qty', '>', 0.0),
-                ('qty_done', '=', 0.0),
                 ('id', '!=', self.id),
             ]
             oudated_candidates = self.env['stock.move.line'].search(oudated_move_lines_domain)
@@ -474,7 +486,10 @@ class StockMoveLine(models.Model):
                 if float_compare(candidate.product_qty, quantity, precision_rounding=rounding) <= 0:
                     quantity -= candidate.product_qty
                     move_to_recompute_state |= candidate.move_id
-                    candidate.unlink()
+                    if candidate.qty_done:
+                        candidate.product_uom_qty = 0.0
+                    else:
+                        candidate.unlink()
                 else:
                     # split this move line and assign the new part to our extra move
                     quantity_split = float_round(
